Fix circular import issue
